feat: Implementar rediseño inicial a tema oscuro moderno (corregido)
Se aplica la primera fase del rediseño visual a un tema oscuro,
moderno y profesional, centrándose en la página de inicio y las
páginas de login de docente y estudiante.

Este commit corrige un error anterior donde el contenido de los archivos
de las páginas de login fue accidentalmente concatenado a App.jsx.
Los archivos han sido restaurados a sus contenidos correctos.

Cambios Principales:

1.  **Estilos Globales (`index.css`, `index.html`):**
    *   Importada fuente Poppins.
    *   Definidas variables CSS para la nueva paleta de colores oscura.
    *   Aplicado fondo de gradiente global (azul oscuro a morado oscuro).
    *   Estilos base para elementos comunes adaptados al tema oscuro.

2.  **Página de Inicio (`/`):**
    *   Layout centrado con título "Theaterboard 2025" estilizado.
    *   Botones de acción ("Docente", "Estudiante", "Regístrate") rediseñados
        con bordes redondeados, colores distintivos e íconos SVG.

3.  **Páginas de Login (Docente y Estudiante):**
    *   Formularios presentados en tarjetas centradas con fondo
        semitransparente.
    *   Inputs y botones de login rediseñados según la nueva guía de estilo,
        incluyendo íconos SVG para los botones de ingreso.
    *   Enlaces secundarios estilizados.

4.  **Header/Footer:**
    *   Estilos adaptados para una mejor integración con el tema oscuro.

Este commit establece la base visual para futuras mejoras y la
extensión del diseño al resto de la aplicación.
